+++ pages/ingredients.php
@@ -115,36 +115,39 @@ $res_ingCategory = mysqli_query($conn, "SELECT id,image,name,notes FROM ingCateg
 
 <!--CSV IMPORT-->
 <div class="modal fade" id="csv_import" tabindex="-1" role="dialog" aria-labelledby="csv_import" aria-hidden="true">
-  <div class="modal-dialog" role="document">
+  <div class="modal-dialog pv-modal-xxl" role="document">
     <div class="modal-content">
       <div class="modal-header">
-        <h5 class="modal-title" id="csv_import">Import ingredients from CSV file</h5>
+        <h5 class="modal-title">Import ingredients from CSV file</h5>
         <button type="button" class="close" data-dismiss="modal" aria-label="Close">
           <span aria-hidden="true">&times;</span>
         </button>
       </div>
       <div class="modal-body">
        <div id="CSVImportMsg"></div>
-		<form method="post" action="javascript:importCSV()" enctype="multipart/form-data" id="csvform">
-       <table width="100%">
-       		<tr>
-    	   	<td width="92" valign="top">CSV File:</td>
-				<td width="1533" colspan="3">
-                	<input type="file" id="ingCSV" name="ingCSV" />
-				</td>
-			</tr>
-		</table>
-         <strong>WARNING:</strong><br />
-      		Make sure your CSV file follows the guidelines as documented <a href="https://www.jbparfum.com/knowledge-base/3-ingredients-import-csv" target="_blank">here</a>
+        <div id=process_area>
+
+           <table width="100%">
+                <tr>
+                <td width="92" valign="top">CSV File:</td>
+                    <td width="1533" colspan="3">
+                        <input type="file" id="CSVFile" name="CSVFile" />
+                    </td>
+                </tr>
+            </table>
+        
+        </div>
+        <div id="step_upload" class="modal-body"></div>
+        <div class="alert alert-info">Select and match the fields in you CSV file, if a column isn't applicable, set it to <strong>None</strong>. Any existing data in your database will not be replaced and or updated if exists in CSV.</div>
       </div>
       <div class="modal-footer">
-        <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-        <input type="submit" name="button" class="btn btn-primary" id="btnImportCSV" value="Import">
+        <input type="button" class="btn btn-secondary" data-dismiss="modal" id="btnCloseCsv" value="Cancel">
+        <input type="submit" class="btn btn-primary" id="btnImportCSV" value="Import">
       </div>
-      </form>
     </div>
   </div>
 </div>  
+
 <!--PV ONLINE IMPORT-->
 <div class="modal fade" id="pv_online_import" tabindex="-1" role="dialog" aria-labelledby="pv_online_import" aria-hidden="true">
   <div class="modal-dialog" role="document">
@@ -332,71 +335,6 @@ $('#pv_online_upload').on('click', '[id*=btnUpload]', function () {
 	  });
 });
 <?php } ?>
-
-function delete_ingredient(id){
-	
-	$.ajax({
-		url: 'pages/update_data.php', 
-		type: 'GET',
-		data: {
-			ingredient: "delete",
-			ing_id: id,
-			},
-		dataType: 'html',
-		success: function (data) {
-		  	$('#innermsg').html(data);
-			list_ingredients();
-		}
-	  });
-};
-
-function importCSV(){
-    $("#CSVImportMsg").html('<div class="alert alert-info alert-dismissible">Please wait, file upload in progress....</div>');
-	$("#btnImport").prop("disabled", true);
-		
-	var fd = new FormData();
-    var files = $('#ingCSV')[0].files;
-        
-       if(files.length > 0 ){
-          fd.append('ingCSV',files[0]);
-
-        $.ajax({
-           url: 'pages/upload.php?type=ingCSVImport',
-           type: 'POST',
-           data: fd,
-           contentType: false,
-           processData: false,
-           success: function(response){
-             if(response != 0){
-               $("#CSVImportMsg").html(response);
-				$("#btnImport").prop("disabled", false);
-              }else{
-                $("#CSVImportMsg").html('<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> File upload failed!</div>');
-				$("#btnImport").prop("disabled", false);
-              }
-            },
-         });
-  }else{
-	$("#CSVImportMsg").html('<div class="alert alert-danger alert-dismissible"><a href="#" class="close" data-dismiss="alert" aria-label="close">x</a><strong>Error:</strong> Please select a file to upload!</div>');
-	$("#btnImport").prop("disabled", false);
-  }
-};
-
-function importING(name) {	  
-	$.ajax({ 
-		url: 'pages/update_data.php', 
-		type: 'GET',
-		data: {
-			'import': 'ingredient',
-			'name': name,
-			},
-		dataType: 'html',
-		success: function (data) {
-			$('#innermsg').html(data);
-		}
-	  });
-};
-
-
-
 </script>
+
+<script src="/js/ingredients.js"></script>