+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 - DEPRECATED: Ingredients card view
 - Add option to print the formula without csv export
 - CSV Formula import improvements
+- CSV Ingredients import improvements
 - TECH Preview: PV Online share With a friend
 
 ### Version 4.7